Allow resource level links to be defined
Graphiti supports links at a relationship level and at the top-level of the
jsonapi document.

Relationship level links are mainly to support lazy loading, and top level links
are used solely for pagination.

This commit adds support for defining links at the resource level.

These links could be used to provide the URL to the resource, or alternative
representation of the resource, or even actions to perform against the resource.

{
  "data": {
    "id": 1,
    "type": "tests",
    "attributes": {},
    "relationships": {},
    "links": {
      "self": "http://test.com/api/v1/tests/1"
    }
  }
}

To define a link on a resource:

link :self do |model|
  "#{self.endpoint[:url]}/#{model.id}"
end